requirements.txt
langchain_google_genai
streamlit
pandas
loguru
python-dotenv

